+++ src/Lab1/Obstacles/Obstacle.cs
@@ -2,12 +2,12 @@ namespace Itmo.ObjectOrientedProgramming.Lab1.Obstacles;
 
 public abstract class Obstacle
 {
-    protected Obstacle(int damage, bool killCrew)
+    protected Obstacle(double damage, bool killCrew)
     {
         Damage = damage;
         KillCrew = killCrew;
     }
 
-    public int Damage { get; }
+    public double Damage { get; }
     public bool KillCrew { get; }
 }
